Minor refactor. Mostly cleanup
