+++ .gitignore
@@ -34,3 +34,4 @@ sysinfo.txt
 # Builds
 *.apk
 *.unitypackage
+Assets/ImportExport/Exports/
