+++ Assets/ImportExport/Examples.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 97dd6f28430e2f94e8ed3f2684f0778e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
