+++ Assets/ImportExport/Examples/CustomLogic.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e156628e9ba82c34b9bcd4150fc2892d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
