+++ Assets/ImportExport/Examples/CustomLogic/QuaternionCustomMappingLogic.cs.meta
