+++ Assets/ImportExport/Models/MergeNode.cs
@@ -1,5 +1,4 @@
 using System;
-using Newtonsoft.Json;
 using UnityEngine;
 
 namespace importerexporter.models
