--- Assets/ImportExport/Models/PrefabTestScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 57720f0225594c843a5837cc7b50f1ed
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
