+++ Assets/ImportExport/Utility/Computation/MD4.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Linq;
+﻿using System.Linq;
 using System.Collections.Generic;
 using System.Security.Cryptography;
 
@@ -174,29 +173,6 @@ namespace importerexporter.utility
             unchecked
             {
                 return ROL(a + (b ^ c ^ d) + xk + 0x6ed9eba1, s);
-            }
-        }
-    }
-
-    public static class FileIDUtil
-    {
-        public static int Compute(Type t)
-        {
-            string toBeHashed = "s\0\0\0" + t.Namespace + t.Name;
-
-            using (HashAlgorithm hash = new MD4())
-            {
-                byte[] hashed = hash.ComputeHash(System.Text.Encoding.UTF8.GetBytes(toBeHashed));
-
-                int result = 0;
-
-                for (int i = 3; i >= 0; --i)
-                {
-                    result <<= 8;
-                    result |= hashed[i];
-                }
-
-                return result;
             }
         }
     }
