+++ Assets/ImportExport/Utility/MainThreadDispatcherEditorWindow.cs
@@ -5,8 +5,8 @@ using UnityEditor;
 /// <summary>
 /// EditorWindow that can run actions on the MainThread for user input 
 /// </summary>
-public class MainThreadDispatcherEditorWindow:EditorWindow {
-
+public class MainThreadDispatcherEditorWindow : EditorWindow
+{
     private static readonly Queue<Action> _executionQueue = new Queue<Action>();
 
     public MainThreadDispatcherEditorWindow()
@@ -14,25 +14,31 @@ public class MainThreadDispatcherEditorWindow:EditorWindow {
         _instance = this;
     }
 
-    public void Update() {
-        lock(_executionQueue) {
-            while (_executionQueue.Count > 0) {
+    public void Update()
+    {
+        lock (_executionQueue)
+        {
+            while (_executionQueue.Count > 0)
+            {
                 _executionQueue.Dequeue().Invoke();
             }
         }
     }
+
     private static MainThreadDispatcherEditorWindow _instance = null;
 
-    public static bool Exists() {
+    public static bool Exists()
+    {
         return _instance != null;
     }
-    
+
     public void Enqueue(Action action)
     {
         _executionQueue.Enqueue(action);
     }
-    public static MainThreadDispatcherEditorWindow Instance() {
+
+    public static MainThreadDispatcherEditorWindow Instance()
+    {
         return _instance;
     }
-
 }
