+++ Assets/ImportExport/Utility/Serialization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7c3944c9c19f0274d9366c63571ee54e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
