--- Assets/scripts/TestScriptQuaternion.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-//using importerexporter.controllers.customlogic;
-
-//[CustomMappingLogic(typeof(QuaternionCustomMappingLogic))]
-public class TestScriptQuaternion : MonoBehaviour
-{
-    public Vector3 testQuaternion;
-    
-}
